requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiofiles==23.2.1
pydantic==2.7.1
pydantic-settings==2.12.0
python-dotenv==1.0.1
redis==5.0.4
rq==1.16.2
httpx==0.25.0
langchain==0.2.5
langchain-openai==0.1.7
pytest==8.2.2
prometheus-fastapi-instrumentator==7.0.0
